% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ballads}
\alias{ballads}
\title{English Broadside Ballad Archive Corpus.}
\format{A character vector with 5818 elements, names are ballad ids.}
\source{
\url{https://ebba.english.ucsb.edu/}
}
\usage{
ballads
}
\description{
A collection of the text of 5818 ballads. Each text is named according to its id in the EBBA catalogue.
Each token delimited by space.
}
\keyword{datasets}
