% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inaugural}
\alias{inaugural}
\title{C-Span Inaugural Address Corpus.}
\format{A character vector with 5818 elements, names are ballad ids.}
\source{
\url{https://archive.org/details/Inaugural-Address-Corpus-1789-2009}
}
\usage{
inaugural
}
\description{
A collection of the texts of US presidential inaugural addresses from 1789-2009.
Each token delimited by space.
}
\keyword{datasets}
