% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moby_dick}
\alias{moby_dick}
\title{Moby-Dick.}
\format{A character vector with 1 element that is the raw text of the book.}
\source{
\url{http://www.gutenberg.org/ebooks/2701}
}
\usage{
moby_dick
}
\description{
Full text of Moby-Dick by Herman Melville from project Gutenberg.
}
\keyword{datasets}
