% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movie_reviews}
\alias{movie_reviews}
\title{Movie Reviews Corpus.}
\format{A character vector with 2000 elements.}
\source{
\url{https://github.com/cstahmer/text_mining_with_r/tree/master/data/movie_reviews}
}
\usage{
movie_reviews
}
\description{
A collection of 2000 movie reviews. Each token is delimited by a space. Set to lowercase.
}
\keyword{datasets}
