% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tweets_corpus}
\alias{tweets_corpus}
\title{Sample Tweets Corpus.}
\format{A dataframe with 13 variables. The $text column contains the raw text of the tweets.}
\usage{
tweets_corpus
}
\description{
A collection of 99,997 tweets and their metadata. These tweets are a random sample of
the data collected by Jiayi Young from the 2016 US presidential election.
}
\keyword{datasets}
