% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_management}
\alias{water_management}
\title{Sample Water Resource Management Literature Review PDF Corpus.}
\format{A character vector with 499 elements.}
\usage{
water_management
}
\description{
A collection of 499 text files containing the output of Abbyy Fine Reader OCR applied to pdfs.
This collection is a sample from a large corpus of pdfs collected in 2018 of research papers
published on water resource management in South America.
}
\keyword{datasets}
