% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zoonomics}
\alias{zoonomics}
\title{Sample of Zoonomics Corpus.}
\format{A character vector with 10 elements.}
\usage{
zoonomics
}
\description{
A collection of just 10 text files containing the output of Tesseract OCR applied to pdfs on zoonomics.
}
\keyword{datasets}
